python/requirements.txt
# Phase II: Bayesian Inference & Visualization

# Core Data Processing
pandas>=2.0
numpy>=1.24
pyarrow>=14.0

# Bayesian Modeling
orbit-ml>=1.1.5
stan>=3.1
arviz>=0.15

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.16
scipy>=1.10

# Utilities
jupyter>=1.0
ipython>=8.0
scikit-learn>=1.3
tqdm>=4.65
python-dotenv>=1.0
